backend/ocr-service/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# OCR and Image Processing
paddlepaddle==3.2.1
paddleocr==2.9.1
opencv-python==4.10.0.84
Pillow==11.0.0

# PDF Processing
pdf2image==1.17.0
PyMuPDF==1.25.1

# Data Processing
numpy==1.26.4
pandas==2.2.3

# Environment and Configuration
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.0

# CORS and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
aiofiles==23.2.1
python-dateutil==2.8.2

# Logging
loguru==0.7.2
